+++ lib/ui/settings/SettingsVideoTutorialPanel.dart
@@ -21,6 +21,7 @@ import 'package:illinois/service/NativeCommunicator.dart';
 import 'package:illinois/ui/widgets/HeaderBar.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/styles.dart';
+import 'package:rokwire_plugin/ui/widgets/triangle_painter.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 import 'package:video_player/video_player.dart';
 
@@ -52,6 +53,7 @@ class _SettingsVideoTutorialPanelState extends State<SettingsVideoTutorialPanel>
     String? tutorialUrl = Config().videoTutorialUrl;
     if (StringUtils.isNotEmpty(tutorialUrl)) {
       _controller = VideoPlayerController.network(tutorialUrl!);
+      _controller!.addListener(_checkVideoEnded);
       _initializeVideoPlayerFuture =
           _controller!.initialize().then((_) => _controller!.play()); // Automatically play video after initialization
     }
@@ -90,7 +92,13 @@ class _SettingsVideoTutorialPanelState extends State<SettingsVideoTutorialPanel>
           future: _initializeVideoPlayerFuture,
           builder: (context, snapshot) {
             if (snapshot.connectionState == ConnectionState.done) {
-              return Center(child: AspectRatio(aspectRatio: _controller!.value.aspectRatio, child: VideoPlayer(_controller!)));
+              return GestureDetector(
+                  onTap: _onTapPlayPause,
+                  child: Center(
+                      child: Stack(alignment: Alignment.center, children: [
+                    AspectRatio(aspectRatio: _controller!.value.aspectRatio, child: VideoPlayer(_controller!)),
+                    _buildPlayButton()
+                  ])));
             } else {
               return const Center(child: CircularProgressIndicator());
             }
@@ -100,5 +108,65 @@ class _SettingsVideoTutorialPanelState extends State<SettingsVideoTutorialPanel>
           child: Text(Localization().getStringEx('panel.settings.video_tutorial.video.missing.msg', 'Missing video'),
               style: TextStyle(color: Styles().colors!.white, fontSize: 20, fontFamily: Styles().fontFamilies!.bold)));
     }
+  }
+
+  Widget _buildPlayButton() {
+    final double buttonWidth = 80;
+    final double buttonHeight = 50;
+    bool buttonVisible = _isPlayerInitialized && !_isPlaying;
+    return Visibility(
+        visible: buttonVisible,
+        child: Container(
+            decoration: BoxDecoration(color: Styles().colors!.iconColor, borderRadius: BorderRadius.all(Radius.circular(10))),
+            width: buttonWidth,
+            height: buttonHeight,
+            child: Center(
+                child: Column(mainAxisAlignment: MainAxisAlignment.center, children: [
+              Container(
+                  width: (buttonHeight / 2),
+                  child: CustomPaint(
+                      painter: TrianglePainter(
+                          painterColor: Styles().colors!.white,
+                          horzDir: TriangleHorzDirection.rightToLeft,
+                          vertDir: TriangleVertDirection.topToBottom),
+                      child: Container(height: (buttonHeight / 4)))),
+              Container(
+                  width: (buttonHeight / 2),
+                  child: CustomPaint(
+                      painter: TrianglePainter(
+                          painterColor: Styles().colors!.white,
+                          horzDir: TriangleHorzDirection.rightToLeft,
+                          vertDir: TriangleVertDirection.bottomToTop),
+                      child: Container(height: (buttonHeight / 4))))
+            ]))));
+  }
+
+  void _onTapPlayPause() {
+    if (!_isPlayerInitialized) {
+      return;
+    }
+    if (_isPlaying) {
+      _controller?.pause();
+    } else {
+      _controller?.play();
+    }
+    setState(() {});
+  }
+
+    void _checkVideoEnded() {
+    if (_controller != null) {
+      bool videoEnded = (_controller!.value.position == _controller!.value.duration);
+      if (videoEnded && mounted) {
+        setState(() {});
+      }
+    }
+  }
+
+  bool get _isPlaying {
+    return (_controller?.value.isPlaying ?? false);
+  }
+
+  bool get _isPlayerInitialized {
+    return (_controller?.value.isInitialized ?? false);
   }
 }