+++ lib/service/Gies.dart
@@ -1,9 +1,10 @@
 import 'package:illinois/service/Storage.dart';
+import 'package:rokwire_plugin/service/groups.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/service.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
-class Gies with Service{
+class Gies with Service implements NotificationsListener{
   static const String notifyPageChanged  = "edu.illinois.rokwire.gies.service.page.changed";
   static const String notifyPageCompleted  = "edu.illinois.rokwire.gies.service.page.completed";
   static const String notifySwipeToPage  = "edu.illinois.rokwire.gies.service.action.swipe.page";
@@ -13,6 +14,7 @@ class Gies with Service{
 
   late Map<int, Set<String>> _progressPages;
   Set<String>? _completedPages;
+  Set<String>? _verifiedPages;
   List<int>? _progressSteps;
 
   // Singletone instance
@@ -26,15 +28,28 @@ class Gies with Service{
 
   // Service
   @override
+  void createService() {
+    super.createService();
+    NotificationService().subscribe(this, [Groups.notifyUserMembershipUpdated, Groups.notifyGroupUpdated, Groups.notifyGroupCreated, Groups.notifyUserGroupsUpdated]);
+    super.createService();
+  }
+
+  @override
+  void destroyService() {
+    NotificationService().unsubscribe(this);
+    super.destroyService();
+  }
+
+  @override
   Future<void> initService() async{
     await super.initService();
     _navigationPages = Storage().giesNavPages ?? [];
     _completedPages = Storage().giesCompletedPages ?? Set<String>();
-
     AppBundle.loadString('assets/gies.json').then((String? assetsContentString) {
         _pages = JsonUtils.decodeList(assetsContentString);
         _buildProgressSteps();
         _ensureNavigationPages();
+        _initialPageVerification();
     });
   }
 
@@ -125,6 +140,7 @@ class Gies with Service{
       if ((pageId != null) && pageId.isNotEmpty && !Gies().completedPages!.contains(pageId)) {
           _completedPages!.add(pageId);
         Storage().giesCompletedPages = _completedPages;
+        _verifyPage(pageId);
         NotificationService().notify(notifyPageCompleted, pageId);
       }
     }
@@ -163,7 +179,8 @@ class Gies with Service{
 
   bool isProgressStepCompleted(int? progressStep) {
     Set<String>? progressPages = _progressPages[progressStep];
-    return (progressPages == null) || _completedPages!.containsAll(progressPages);
+    return (progressPages == null) ||
+        (_completedPages!.containsAll(progressPages) && (_verifiedPages?.containsAll(progressPages) ?? false));
   }
 
   String? setCurrentNotes(List<dynamic>? notes, String? pageId) {
@@ -190,12 +207,64 @@ class Gies with Service{
     return currentPageId;
   }
 
+  void _initialPageVerification(){
+    _verifiedPages = new Set();
+    _loadPageVerification();
+  }
+
+  void _loadPageVerification({bool notify = false}){
+    completedPages?.forEach((pageId) {
+        _verifyPage(pageId, notify: notify);
+    });
+  }
+
+  void _verifyPage(String? page, {bool notify = false}){
+    if(page == null || _verifiedPages == null) {
+      return;
+    }
+
+    if(_isPageGroupMembershipApproved(page)){
+      if(!_verifiedPages!.contains(page)) {
+        _verifiedPages?.add(page);
+        if(notify){
+          NotificationService().notify(notifyPageCompleted);
+        }
+      }
+    } else {
+      if(_verifiedPages!.contains(page)){
+        _verifiedPages!.remove(page);
+        if(notify){
+          NotificationService().notify(notifyPageCompleted);
+        }
+      }
+    }
+  }
+
+  bool _isPageGroupMembershipApproved(String? pageId){
+    if(StringUtils.isEmpty(pageId))
+      return false;
+
+    dynamic pageData = getPage(id: pageId);
+    String? groupName = pageData is Map ?  JsonUtils.stringValue(pageData["group_name"]) : null;
+    Set<String>? groupsNames = Groups().userGroupNames;
+    return groupName != null &&
+        (groupsNames?.contains(groupName) ?? false);
+  }
+
+  bool isPageVerified(String? pageId){
+    return pageId!= null && (_verifiedPages?.contains(pageId) ?? false);
+  }
+
   List<dynamic>? get pages{
     return _pages;
   }
 
   Set<String>? get completedPages{
     return _completedPages;
+  }
+
+  Set<String>? get verifiedPages{
+    return _verifiedPages;
   }
 
   List<dynamic>? get navigationPages{
@@ -250,13 +319,21 @@ class Gies with Service{
   //Utils
   bool _pageCanComplete(Map? page) {
     List<dynamic>? buttons = (page != null) ? JsonUtils.listValue(page['buttons']) : null;
+    List<dynamic>? bnavigationButtons = (page != null) ? JsonUtils.listValue(page['navigation_buttons']) : null;
     if (buttons != null) {
       for (dynamic button in buttons) {
         if ((button is Map) && pageButtonCompletes(button)) {
           return true;
         }
       }
+    } else if (bnavigationButtons!=null){
+      for (dynamic button in bnavigationButtons) {
+        if ((button is Map) && pageButtonCompletes(button)) {
+          return true;
+        }
+      }
     }
+
     return false;
   }
 
@@ -266,5 +343,15 @@ class Gies with Service{
 
   int? getPageProgress(Map<String, dynamic>? page) {
     return (page != null) ? (JsonUtils.intValue(page['progress']) ?? JsonUtils.intValue(page['progress-possition'])) : null;
+  }
+
+  @override
+  void onNotification(String name, param) {
+     if (name == Groups.notifyUserMembershipUpdated ||
+         name == Groups.notifyGroupUpdated ||
+         // name == Groups.notifyGroupCreated ||
+         name == Groups.notifyUserGroupsUpdated) {
+       _loadPageVerification(notify: true);
+     }
   }
 }