+++ lib/ui/home/HomeGiesWidget.dart
@@ -28,8 +28,7 @@ class _HomeGiesWidgetState extends State<HomeGiesWidget> implements Notification
   @override
   void initState() {
     super.initState();
-    NotificationService().subscribe(this, [Gies.notifyPageChanged]);
-
+    NotificationService().subscribe(this, [Gies.notifyPageChanged, Gies.notifyPageCompleted]);
   }
 
   @override
@@ -258,7 +257,7 @@ class _HomeGiesWidgetState extends State<HomeGiesWidget> implements Notification
 
   @override
   void onNotification(String name, param) {
-    if(name == Gies.notifyPageChanged){
+    if(name == Gies.notifyPageChanged || name == Gies.notifyPageCompleted){
       setState(() {});
     }
   }