+++ CHANGELOG.md
@@ -5,15 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [4.0.3] - 2022-05-03
 ### Added
 - "CC" button in video panels [#1602](https://github.com/rokwire/illinois-app/issues/1602).
+- Gies onboarding updates [#1605](https://github.com/rokwire/illinois-app/issues/1605).
 
 ## [4.0.2] - 2022-05-03
 ### Changed
 - "Health Illini Podcast" renamed to "Healthy Illini Podcast" [#1485](https://github.com/rokwire/illinois-app/issues/1485).
 ### Added
 - "Play" button in video panels [#1602](https://github.com/rokwire/illinois-app/issues/1602).
-- Gies onboarding updates [#1605](https://github.com/rokwire/illinois-app/issues/1605).
 
 ## [4.0.1] - 2022-04-29
 ### Added